refactor: organized the code, but keeping the same functioning
