+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "bcrypt": "^5.0.1",
     "cors": "^2.8.5",
+    "dayjs": "^1.11.2",
     "dotenv": "^16.0.0",
     "express": "^4.18.1",
     "joi": "^17.6.0",
