+++ src/routes/authRoutes.js
@@ -1,9 +1,10 @@
 import { Router } from "express"
 
-import { signup } from "../controllers/authController.js"
+import { signUp } from "../controllers/authController.js"
+import { signUpValidation } from "../middlewares/authValidationMiddleware.js"
 
 const userRouter = Router()
 
-userRouter.post("/sign-up", signup)
+userRouter.post("/sign-up", signUpValidation, signUp)
 
 export default userRouter
